# src/main/java/com/example/bankcards/dto/card/ChangeStatusRequest.java
package com.example.bankcards.dto.card;

import com.example.bankcards.entity.CardStatus;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class ChangeStatusRequest {

    @NotNull
    private CardStatus status;
}